+++ src/ui-component/editor/ajsFlow/nodes/JobGroupNode.tsx
@@ -5,19 +5,23 @@ import DescriptionIcon from '@mui/icons-material/Description';
 import { G } from "../../../../domain/models/units/G";
 import { ActionIcon, AjsNode, nodeActionsSxProps, nodeSxProps, NameOrComment, TyTitle } from "./AjsNode";
 import { handleClickDialogOpen, handleKeyDownDialogOpen } from "./Utils";
+import classNames from "classnames";
 
 export type JobGroupNode = Node<AjsNode<G>, 'jobgroup'>;
 type JobGroupNodeProp = NodeProps<JobGroupNode>;
 const JobGroupNode: FC<JobGroupNodeProp> = ({ data }: JobGroupNodeProp) => {
 
     console.log('render JobGroupNode.');
-    const { unitEntity } = data;
+    const { unitEntity, isAncestor } = data;
 
     return (
         <>
             <Stack
                 id={unitEntity.id}
                 sx={nodeSxProps}
+                className={classNames({
+                    ancestor: isAncestor,
+                })}
             >
                 <TyTitle ty={unitEntity.ty.value()} gty={unitEntity.gty?.value() as 'n' | 'p'} />
                 {/* action */}